style(util): Integrate all function of captcha into captcha.go
Closes: #24
